requirements.txt
--extra-index-url=https://download.pytorch.org/whl/cu121
streamlit==1.34.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
PyYAML==6.0.1
open_clip_torch==2.24.0
black==24.4.2
hydra-core==1.3.2
